Updates per PR feedback - rename BitwiseOperator; Formatting, organization, etc...
