+++ src/AndcultureCode.CSharp.Core/Enumerations/HttpVerb.cs
@@ -8,8 +8,8 @@ namespace AndcultureCode.CSharp.Core.Enumerations
     public enum HttpVerb
     {
         GET = 0,
+        DELETE = 3,
         HEAD = 1,
-        PUT = 2,
-        DELETE = 3
+        PUT = 2
     }
 }
