+++ src/AndcultureCode.CSharp.Core/Enumerations/LogicalOperator.cs
@@ -3,7 +3,7 @@ namespace AndcultureCode.CSharp.Core.Enumerations
     /// <summary>
     /// Simple Bitwise enumeration
     /// </summary>
-    public enum BitwiseOperator
+    public enum LogicalOperator
     {
         And = 1,
         Or = 2
